Add files via upload
This is an AirBnB Project in Tableau that shows the  states in United States of America (USA) and the average amount of money made in a month for an AirBnB apartment depending on the number of bedrooms. It also indicates the best time of the year most profit is made in AirBnB for an investor.